search rom function modified

+++ Programs/Aufgabe4/Inc/one_wire_bus_functions.h
@@ -60,7 +60,7 @@ void readROMNumber(uint8_t rom[8]);
 * which allows the master to determine the number of 
 * slaves and their device types
 */
-uint8_t searchROM(uint8_t *rom, uint8_t last_device_flag);
+int searchROM(uint8_t *rom, int last_device_flag);
 
 void initOnewire();
 
@@ -68,6 +68,6 @@ void initOnewire();
 int skip_rom();
 int match_rom(uint8_t rom[8]);
 
-void detect_sensors(uint8_t roms[][8], uint8_t count);
+void detect_sensors(uint8_t roms[][8], int count);
 
 #endif
